requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.3
sqlalchemy==2.0.23
selenium==4.15.2
fake-useragent==1.4.0
pydantic==2.5.0
tenacity==8.2.3
ratelimit==2.2.1
scrapy==2.11.0
schedule==1.2.0
python-dateutil==2.8.2
tabulate==0.9.0
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
